requirements.txt
streamlit==1.39.0
numpy==2.1.2
Pillow==10.4.0
folium==0.17.0
streamlit-folium==0.22.1
pyarrow==17.0.0
requests==2.32.3
reportlab==4.2.2
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0